Move some code from selector to a separate widget
