+++ src/ui/widgets/mod.rs
@@ -2,10 +2,12 @@ pub use self::button::*;
 pub use self::buttons_group::*;
 pub use self::dialog::*;
 pub use self::input::*;
-pub use self::selector::*;
+pub use self::select::*;
+pub use self::side_select::*;
 
 mod button;
 mod buttons_group;
 mod dialog;
 mod input;
-mod selector;
+mod select;
+mod side_select;
